abkhazia_blf_cuts/requirements.txt
# requirements.txt

altgraph==0.17.4
annotated-types==0.6.0
asttokens==2.4.1
auto-py-to-exe==2.42.0
bottle==0.12.25
bottle-websocket==0.2.9
canlib==1.24.935
cffi==1.16.0
colorama==0.4.6
Eel==0.16.0
executing==2.0.1
future==0.18.3
gevent==23.9.1
gevent-websocket==0.10.1
greenlet==3.0.3
icecream==2.1.3
packaging==23.2
pefile==2023.2.7
pycparser==2.21
pydantic==2.6.0
pydantic_core==2.16.1
Pygments==2.17.2
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2024.0
pyparsing==3.1.1
python-can==4.3.1
pywin32==306
pywin32-ctypes==0.2.2
setuptools==69.0.3
six==1.16.0
tk==0.1.0
typing_extensions==4.9.0
whichcraft==0.6.1
wrapt==1.16.0
zope.event==5.0
zope.interface==6.1
distlib==0.3.8
filelock==3.13.1
greenlet==3.0.3
msgpack==1.0.7
platformdirs==4.1.0
pynvim==0.5.0
virtualenv==20.25.0
altgraph==0.17.4
annotated-types==0.6.0
asttokens==2.4.1
auto-py-to-exe==2.42.0
bottle==0.12.25
bottle-websocket==0.2.9
canlib==1.24.935
cffi==1.16.0
colorama==0.4.6
Eel==0.16.0
executing==2.0.1
future==0.18.3
gevent==23.9.1
gevent-websocket==0.10.1
greenlet==3.0.3
icecream==2.1.3
Nuitka==2.0.5
ordered-set==4.1.0
packaging==23.2
pefile==2023.2.7
pycparser==2.21
pydantic==2.6.0
pydantic_core==2.16.1
Pygments==2.17.2
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2024.0
pyparsing==3.1.1
python-can==4.3.1
pywin32==306
pywin32-ctypes==0.2.2
setuptools==69.0.3
six==1.16.0
tk==0.1.0
typing_extensions==4.9.0
whichcraft==0.6.1
wrapt==1.16.0
zope.event==5.0
zope.interface==6.1
zstandard==0.22.0
